fix: separate old and new scripts with version
