+++ src/v1/token-transfer.ts
@@ -8,9 +8,9 @@ import {
   exitSuccess,
   printDivider,
   MICRO_UNITS,
-} from "../lib/utils";
-import { getNonce, STACKS_NETWORK } from "../lib/stacks";
-import { getCCBalance } from "../lib/citycoins";
+} from "../../lib/utils";
+import { getNonce, STACKS_NETWORK } from "../../lib/stacks";
+import { getCCBalance } from "../../lib/citycoins";
 import {
   bufferCVFromString,
   noneCV,
